+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow [Semantic Versions](https://semver.org/).
 ## Version 0.2.2 (unreleased)
 
 - Update dependencies and setup.cfg
+- Add navigation links to the previous and the next post
 
 
 ## Version 0.2.1